Revert "Create new internal/planutil package (#156)"
This reverts commit dfff363806037ef34ffb237f8e58be13940aaa27.
