+++ internal/overlord/servstate/manager_test.go
@@ -41,7 +41,6 @@ import (
 	"github.com/canonical/pebble/internal/overlord/restart"
 	"github.com/canonical/pebble/internal/overlord/servstate"
 	"github.com/canonical/pebble/internal/overlord/state"
-	"github.com/canonical/pebble/internal/planutil"
 	"github.com/canonical/pebble/internal/testutil"
 	"github.com/canonical/pebble/plan"
 )
@@ -586,7 +585,7 @@ services:
 }
 
 func parseLayer(c *C, order int, label, layerYAML string) *plan.Layer {
-	layer, err := planutil.ParseLayer(order, label, []byte(layerYAML))
+	layer, err := plan.ParseLayer(order, label, []byte(layerYAML))
 	c.Assert(err, IsNil)
 	return layer
 }
