+++ plan/suite_test.go
@@ -1,3 +1,4 @@
+//
 // Copyright (c) 2020 Canonical Ltd
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -12,7 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package planutil_test
+package plan_test
 
 import (
 	"testing"
